EDITOR-#217 [Frontend] re-route
