+++ frontend/src/containers/MaterialPage.tsx
@@ -24,9 +24,9 @@ import type { MaterialInput } from "@/shared/type";
 import MaterialImportButton from "@/components/MaterialAndTool/MaterialImportButton";
 import { useToast } from "@/components/ui/use-toast";
 import Searchbar from "@/components/Searchbar";
-import { useUser } from "@/context/UserContext";
+import { useUser } from "@/contexts/UserContext";
 import { useSearchParams } from "react-router-dom";
-import { useDebounce } from "@/lib/useDebounce";
+import { useDebounce } from "@/hooks/useDebounce";
 import SkeletonList from "@/components/SkeletonList";
 import { Cpu } from "lucide-react";
 import { useTranslation } from "react-i18next";
