+++ frontend/src/containers/ToolPage.tsx
@@ -21,12 +21,12 @@ import { Textarea } from "@/components/ui/textarea";
 import { Button } from "@/components/ui/button";
 import type { ToolInput } from "@/shared/type";
 import { useToast } from "@/components/ui/use-toast";
-import { useUser } from "@/context/UserContext";
+import { useUser } from "@/contexts/UserContext";
 import ToolList from "@/components/MaterialAndTool/ToolList";
 import ToolImportButton from "@/components/MaterialAndTool/ToolImportButton";
 import Searchbar from "@/components/Searchbar";
 import { useSearchParams } from "react-router-dom";
-import { useDebounce } from "@/lib/useDebounce";
+import { useDebounce } from "@/hooks/useDebounce";
 import SkeletonList from "@/components/SkeletonList";
 import { Hammer } from "lucide-react";
 import { useTranslation } from "react-i18next";
