+++ frontend/src/contexts/WindowContext.tsx
